notebook_to_cloud/ml_deploy_demo/requirements.txt
# production
pyyaml
numpy
pandas
scikit-learn
flask
gunicorn
tensorflow
keras
scipy
imageio

# dev
jupyterlab
coloredlogs
black
flake8

# etc
